Forward mostly done
Hopefully, everything is correct so far. I might be off for An by 1, but such is the curse of being with CSC's.

Back prop is left to do which will take a ton of time to do, probably about 2 hours. We also have to just combine the forward and back prop for the train and predict functions.